refactors transaction generation into another saga
