+++ app/modules/tx/sender/actions.ts
@@ -1,13 +1,13 @@
 import { createAction, createSimpleAction } from "../../actionsUtils";
 import { ITxData } from "./../../../lib/web3/Web3Manager";
-import { ETxSenderType, ETokenType } from './reducer';
+import { ETokenType, ETxSenderType } from "./reducer";
 
 export const txSenderActions = {
   // Modal related actions
   txSenderShowModal: (type: ETxSenderType) => createAction("TX_SENDER_SHOW_MODAL", { type }),
   txSenderHideModal: () => createSimpleAction("TX_SENDER_HIDE_MODAL"),
   // User awaiting actions
-  txSenderAcceptDraft: (txData: ITxData) => createAction("TX_SENDER_ACCEPT_DRAFT", txData),
+  txSenderAcceptDraft: (txData?: ITxData) => createAction("TX_SENDER_ACCEPT_DRAFT", txData!),
   txSenderAccept: () => createSimpleAction("TX_SENDER_ACCEPT"),
   // Signer actions
   txSenderSigned: (txHash: string) => createAction("TX_SENDER_SIGNED", { txHash }),
@@ -22,7 +22,7 @@ export const txSenderActions = {
   // Error Actions
   txSenderError: (error: string) => createAction("TX_SENDER_ERROR", { error }),
 
-  //Transaction flows 
+  //Transaction flows
   startWithdrawEth: () => createSimpleAction("TX_SENDER_START_WITHDRAW_ETH"),
   startUpgrade: (tokenType: ETokenType) => createAction("TX_SENDER_START_UPGRADE", tokenType),
   startInvestment: () => createSimpleAction("TX_SENDER_START_INVESTMENT"),
