+++ app/modules/tx/sender/reducer.ts
@@ -35,6 +35,7 @@ export interface ITxSenderState {
 
 const initialState: ITxSenderState = {
   state: "UNINITIALIZED",
+  // txDetails: { to: "", value: "", gas: "", gasPrice: "", from: "" },
 };
 
 export const txSenderReducer: AppReducer<ITxSenderState> = (
